scripts/db_utils.py
import ee
import os
from dotenv import load_dotenv

def get_car_geometry_mock(car_code):
    """
    Returns a mock ee.Geometry for the specific test CAR.
    In a real scenario, this would query the PostGIS database.
    """
    # Specific test CAR provided by the user
    test_car = "PA-1504802-FCEA8FAD347340D8BD6D3143A9623468"
    
    if car_code == test_car:
        # Coordinates extracted from the user's WKT
        coords = [
            [-54.0292538888889, -1.70762083333333], [-54.0299021548834, -1.70194216654557], 
            [-54.0275752777778, -1.70235416666667], [-54.0275972222222, -1.69943916666667], 
            [-54.0275925, -1.69583083333333], [-54.0276033333333, -1.69221194444444], 
            [-54.0257080555556, -1.69208944444444], [-54.0255525, -1.69207416666667], 
            [-54.0231763888889, -1.69190111111111], [-54.0194411111111, -1.69179805555556], 
            [-54.0090066666667, -1.69362472222222], [-54.0098836111111, -1.69616861111111], 
            [-54.0106744444444, -1.69956305555556], [-54.0114675, -1.70224833333333], 
            [-54.0127169444444, -1.70497916666667], [-54.0042388888889, -1.70369222222222], 
            [-54.0042344444444, -1.70659777777778], [-54.0140338888889, -1.70784527777778], 
            [-54.0207391666667, -1.70373555555556], [-54.0208844444444, -1.70345444444444], 
            [-54.0221402777778, -1.70134972222222], [-54.0292538888889, -1.70762083333333]
        ]
        return ee.Geometry.Polygon([coords])
    else:
        raise ValueError(f"CAR code {car_code} not found in mock data.")

def get_car_geometry_db(car_code):
    """
    Connects to the database and retrieves geometry for the given CAR.
    (To be implemented when DB access is configured)
    """
    # Example structure using psycopg2 (not installed yet likely)
    # import psycopg2
    # conn = psycopg2.connect(...)
    # cur = conn.cursor()
    # cur.execute("SELECT ST_AsText(geom) FROM ngeo.tb_sicar WHERE codigo_car = %s", (car_code,))
    # wkt = cur.fetchone()[0]
    # return ee.Geometry.fromWKT(wkt)
    return get_car_geometry_mock(car_code)
