requirements.txt
earthengine-api
requests
python-dotenv
customtkinter